Move oneway strategies to `hypothesis_helpers.py`
